refactor parl_io for only ISR, return error's from swap
