+++ build-all.sh
@@ -9,19 +9,10 @@ aliases=(
     build-32-d
     build-32h
     build-32r
-    build-c5-l
-    build-c5-bl
-    build-c5-li
+    # build-c5-li
     build-c5-bli
-    build-c6
-    build-c6-bp
-    build-c6-l
-    build-c6-bl
-    build-c6h
     build-c6r
-    build-c6ri
     build-c6-bpi
-    build-c6-li
     build-c6-bli
     build-s3
     build-s3-bp
