--- examples/hello_parl_io.rs
@@ -1,94 +1,0 @@
-//! Example for using the PARL_IO driver with a HUB75 LED matrix display
-//!
-//! This example demonstrates how to use the PARL_IO driver to drive a HUB75 LED
-//! matrix display. It uses the ESP32's Parallel I/O peripheral to send data to
-//! the display.
-
-#![no_std]
-#![no_main]
-#![allow(clippy::uninlined_format_args)]
-
-#[cfg(feature = "defmt")]
-use defmt_rtt as _;
-use embedded_graphics::geometry::Point;
-use embedded_graphics::mono_font::ascii::FONT_5X7;
-use embedded_graphics::mono_font::MonoTextStyleBuilder;
-use embedded_graphics::prelude::RgbColor;
-use embedded_graphics::text::Alignment;
-use embedded_graphics::text::Text;
-use embedded_graphics::Drawable;
-use esp_backtrace as _;
-use esp_hal::clock::CpuClock;
-use esp_hal::gpio::Pin;
-use esp_hal::main;
-use esp_hal::time::Rate;
-use esp_hub75::framebuffer::bitplane::plain::DmaFrameBuffer;
-use esp_hub75::framebuffer::compute_rows;
-use esp_hub75::Color;
-use esp_hub75::Hub75;
-use esp_hub75::Hub75Pins16;
-
-esp_bootloader_esp_idf::esp_app_desc!();
-
-const ROWS: usize = 64;
-const COLS: usize = 64;
-const NROWS: usize = compute_rows(ROWS);
-#[cfg(feature = "esp32c5")]
-const PLANES: usize = 7;
-#[cfg(not(feature = "esp32c5"))]
-const PLANES: usize = 4;
-
-type FBType = DmaFrameBuffer<NROWS, COLS, PLANES>;
-
-#[main]
-fn main() -> ! {
-    let peripherals = esp_hal::init(esp_hal::Config::default().with_cpu_clock(CpuClock::max()));
-
-    let tx_descriptors = esp_hub75::hub75_dma_descriptors!(FBType);
-
-    let pins = Hub75Pins16 {
-        red1: peripherals.GPIO19.degrade(),
-        grn1: peripherals.GPIO20.degrade(),
-        blu1: peripherals.GPIO21.degrade(),
-        red2: peripherals.GPIO22.degrade(),
-        grn2: peripherals.GPIO23.degrade(),
-        blu2: peripherals.GPIO15.degrade(),
-        addr0: peripherals.GPIO10.degrade(),
-        addr1: peripherals.GPIO8.degrade(),
-        addr2: peripherals.GPIO1.degrade(),
-        addr3: peripherals.GPIO0.degrade(),
-        addr4: peripherals.GPIO11.degrade(),
-        blank: peripherals.GPIO5.degrade(),
-        clock: peripherals.GPIO7.degrade(),
-        latch: peripherals.GPIO6.degrade(),
-    };
-
-    let mut hub75 = Hub75::new_async(
-        peripherals.PARL_IO,
-        pins,
-        peripherals.DMA_CH0,
-        tx_descriptors,
-        Rate::from_mhz(20),
-    )
-    .expect("failed to create Hub75!");
-
-    let mut fb = FBType::new();
-    let text_style = MonoTextStyleBuilder::new()
-        .font(&FONT_5X7)
-        .text_color(Color::YELLOW)
-        .background_color(Color::BLACK)
-        .build();
-    let point = Point::new(32, 31);
-    Text::with_alignment("Hello, World!", point, text_style, Alignment::Center)
-        .draw(&mut fb)
-        .expect("failed to draw text");
-    loop {
-        let xfer = hub75
-            .render(&fb)
-            .map_err(|(e, _hub75)| e)
-            .expect("failed to start render!");
-        let (result, new_hub75) = xfer.wait();
-        hub75 = new_hub75;
-        result.expect("transfer failed");
-    }
-}
