--- examples/rustacean_parl_io_isr.rs
@@ -1,105 +1,0 @@
-//! ISR-driven example rendering a Rustacean PNG on a HUB75 LED matrix.
-//!
-//! Unlike the async `rustacean_parl_io` example, this one uses the
-//! `Hub75Isr` driver which manages display refresh entirely from the
-//! PARL_IO `TxEof` interrupt — no Embassy, no async runtime required.
-
-#![no_std]
-#![no_main]
-#![allow(clippy::uninlined_format_args)]
-
-#[cfg(feature = "defmt")]
-use defmt_rtt as _;
-use embedded_graphics::geometry::Point;
-use embedded_graphics::mono_font::ascii::FONT_5X7;
-use embedded_graphics::mono_font::MonoTextStyleBuilder;
-use embedded_graphics::prelude::RgbColor;
-use embedded_graphics::text::Alignment;
-use embedded_graphics::text::Text;
-use embedded_graphics::Drawable;
-use embedded_sprites::{image::Image, include_image, sprite::Sprite};
-use esp_hub75::Color;
-
-use esp_backtrace as _;
-use esp_hal::clock::CpuClock;
-use esp_hal::gpio::Pin;
-use esp_hal::main;
-use esp_hal::time::Rate;
-use esp_hub75::framebuffer::bitplane::plain::DmaFrameBuffer;
-use esp_hub75::framebuffer::compute_rows;
-use esp_hub75::parl_io_isr::Hub75Isr;
-use esp_hub75::Hub75Pins16;
-
-esp_bootloader_esp_idf::esp_app_desc!();
-
-macro_rules! mk_static {
-    ($t:ty,$val:expr) => {{
-        static STATIC_CELL: static_cell::StaticCell<$t> = static_cell::StaticCell::new();
-        #[deny(unused_attributes)]
-        let x = STATIC_CELL.uninit().write($val);
-        x
-    }};
-}
-
-const ROWS: usize = 64;
-const COLS: usize = 64;
-const NROWS: usize = compute_rows(ROWS);
-const PLANES: usize = 7;
-
-type FBType = DmaFrameBuffer<NROWS, COLS, PLANES>;
-
-#[include_image]
-const GRASS_DATA: Image<hub75_framebuffer::Color> = "./images/rustacean-orig-noshadow-64x64.png";
-
-#[main]
-fn main() -> ! {
-    let peripherals = esp_hal::init(esp_hal::Config::default().with_cpu_clock(CpuClock::max()));
-
-    let tx_descriptors = esp_hub75::hub75_dma_descriptors!(FBType);
-
-    let pins = Hub75Pins16 {
-        red1: peripherals.GPIO19.degrade(),
-        grn1: peripherals.GPIO20.degrade(),
-        blu1: peripherals.GPIO21.degrade(),
-        red2: peripherals.GPIO22.degrade(),
-        grn2: peripherals.GPIO23.degrade(),
-        blu2: peripherals.GPIO15.degrade(),
-        addr0: peripherals.GPIO10.degrade(),
-        addr1: peripherals.GPIO8.degrade(),
-        addr2: peripherals.GPIO1.degrade(),
-        addr3: peripherals.GPIO0.degrade(),
-        addr4: peripherals.GPIO11.degrade(),
-        blank: peripherals.GPIO5.degrade(),
-        clock: peripherals.GPIO7.degrade(),
-        latch: peripherals.GPIO6.degrade(),
-    };
-
-    let fb = mk_static!(FBType, FBType::new());
-
-    let rustacean = Sprite::new(Point::new(0, 0), &GRASS_DATA);
-    rustacean.draw(fb).expect("failed to draw image");
-
-    let text_style = MonoTextStyleBuilder::new()
-        .font(&FONT_5X7)
-        .text_color(Color::WHITE)
-        .background_color(Color::BLACK)
-        .build();
-    Text::with_alignment("ISR driven!", Point::new(31, 55), text_style, Alignment::Center)
-        .draw(fb)
-        .expect("failed to draw text");
-
-    let _hub75 = Hub75Isr::new(
-        peripherals.PARL_IO,
-        pins,
-        peripherals.DMA_CH0,
-        tx_descriptors,
-        Rate::from_mhz(20),
-        &*fb,
-    )
-    .expect("failed to create Hub75Isr");
-
-    // Display refreshes automatically via ISR — nothing else to do.
-    loop {
-        core::hint::spin_loop();
-    }
-}
